Add cutsom image stack volume processor in base module
This processor is a modified version of ImageStackVolumeSource processor. It
takes the path of a Whole Slide Image (WSI), extract and stack multiple images
of a particular region of given WSI to construct a 3D volume data. Main motive
is to extract data of small regions at a time of very high dimension WSIs and
to convert them to get 3D volume data.

This processor uses openslide library which can extract any particular region of
a WSI image, given the coordinate of the the top left corner of the region,
width, height and level. Different levels of a WSI contains different dimensions
of slides.

Inport(optional):
-> Coordinates of a particular rectangular region of WSI. If no inport data is
   given then default values are used.

Outport:
-> 3D Volume data

Default Values:
-> Coordinates = (0,0)
-> Width = 1920
-> Height = 1080
-> Level = WSI max level

Features:
-> Slider to change levels (Zoom in/ Zoom out).
-> Slider to change coordinates.